update README to reflect current state
As of early 2019, we've been on CiviDesk for a number of years, and this repo reflects the state as of us moving off WPEngine before the move to CiviDesk.  Soon this will be changing to Pantheon hosting.  Unclear if we're going to continue to use this repo or not.

Updating this README as I promised to do when Dan Chase came on board a few months ago.